fix: align moneyflow volume fields to integer semantics

+++ src/foundation/models/core/equity_moneyflow.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import date
 from decimal import Decimal
 
-from sqlalchemy import Date, Index, Numeric, String
+from sqlalchemy import BigInteger, Date, Index, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src.foundation.models.base import Base, TimestampMixin
@@ -19,21 +19,21 @@ class EquityMoneyflow(TimestampMixin, Base):
 
     ts_code: Mapped[str] = mapped_column(String(16), primary_key=True)
     trade_date: Mapped[date] = mapped_column(Date, primary_key=True)
-    buy_sm_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    buy_sm_vol: Mapped[int | None] = mapped_column(BigInteger)
     buy_sm_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
-    sell_sm_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    sell_sm_vol: Mapped[int | None] = mapped_column(BigInteger)
     sell_sm_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
-    buy_md_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    buy_md_vol: Mapped[int | None] = mapped_column(BigInteger)
     buy_md_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
-    sell_md_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    sell_md_vol: Mapped[int | None] = mapped_column(BigInteger)
     sell_md_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
-    buy_lg_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    buy_lg_vol: Mapped[int | None] = mapped_column(BigInteger)
     buy_lg_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
-    sell_lg_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    sell_lg_vol: Mapped[int | None] = mapped_column(BigInteger)
     sell_lg_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
-    buy_elg_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    buy_elg_vol: Mapped[int | None] = mapped_column(BigInteger)
     buy_elg_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
-    sell_elg_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    sell_elg_vol: Mapped[int | None] = mapped_column(BigInteger)
     sell_elg_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
-    net_mf_vol: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    net_mf_vol: Mapped[int | None] = mapped_column(BigInteger)
     net_mf_amount: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
